Some changes to the database and a few new classes for the ease of use of the database
1. Created sql dml file and a bit of insertion code
2. Got rid of the "user_photos" table from the database. Now a user can have eather one or zero photos
3. Added a new class for storing data - Answer + documentation
4. Also made some changes to the Question class and provided documentation
5. Minor changes to "PhotoAble" and "User" classes
6. Added a new major class "UserDAO" (Data Access Object) for communicating with the database regarding User data. Class is not yet fully implemented but no compile errors will occure. Documentation is present for the present code

+++ src/model/User.java
@@ -40,7 +40,7 @@ public class User extends PhotoAble implements Comparable {
 		this.ID = ID;
 	}
 
-	public long getID() {
+	public int getID() {
 		return ID;
 	}
